+++ lib/ui/screens/auth/join_screen/join_form_view_model.dart
@@ -10,6 +10,7 @@ class JoinFormModel {
   DateTime? userBirth;
   String? userGender;
 
+
   JoinFormModel(
         {required this.userId,
          required this.userPassword,
