+++ lib/ui/screens/auth/login_screen/login_body/login_body.dart
@@ -4,7 +4,7 @@ import 'package:flutter_blog/_core/constants/font.dart';
 import 'package:flutter_blog/_core/constants/move.dart';
 import 'package:flutter_blog/_core/constants/size.dart';
 import 'package:flutter_blog/ui/screens/auth/login_screen/widgets/login_form.dart';
-import 'package:flutter_blog/ui/widgets/custom_logo.dart';
+import 'package:flutter_blog/ui/widgets/icons_and_images/custom_logo.dart';
 import 'package:flutter_blog/ui/widgets/button_items/button/custom_text_button.dart';
 import 'package:flutter_blog/ui/widgets/custom_nav_appbar.dart';
