+++ lib/ui/screens/auth/user_search/id_search_complete_screen/id_search_complete_body.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_blog/_core/constants/font.dart';
 import 'package:flutter_blog/_core/constants/size.dart';
 import 'package:flutter_blog/ui/widgets/button_items/button/custom_elavated_button.dart';
-import 'package:flutter_blog/ui/widgets/custom_svg_image.dart';
+import 'package:flutter_blog/ui/widgets/icons_and_images/custom_svg_image.dart';
 import 'package:flutter_blog/ui/widgets/text_items/custom_text_item.dart';
 
 class IdSearchCompleteBody extends StatelessWidget {
