+++ lib/ui/screens/auth/user_search/pw_search_complete_screen/pw_search_complete_screen.dart
@@ -3,7 +3,7 @@ import 'package:flutter_blog/_core/constants/color.dart';
 import 'package:flutter_blog/_core/constants/font.dart';
 import 'package:flutter_blog/_core/constants/move.dart';
 import 'package:flutter_blog/_core/constants/size.dart';
-import 'package:flutter_blog/ui/widgets/custom_svg_image.dart';
+import 'package:flutter_blog/ui/widgets/icons_and_images/custom_svg_image.dart';
 import 'package:flutter_blog/ui/widgets/text_items/custom_text_item.dart';
 import 'package:flutter_svg/flutter_svg.dart';
