+++ lib/ui/screens/auth/user_search/pw_search_screen/pw_new_set_screen/pw_new_set_body.dart
@@ -9,7 +9,7 @@ import 'package:flutter_blog/ui/screens/auth/user_search/widget/pw_new_form.dart
 import 'package:flutter_blog/ui/widgets/button_items/button/custom_elavated_button.dart';
 import 'package:flutter_blog/ui/widgets/button_items/custom_radio_button_item.dart';
 import 'package:flutter_blog/ui/widgets/custom_date_picker.dart';
-import 'package:flutter_blog/ui/widgets/custom_logo.dart';
+import 'package:flutter_blog/ui/widgets/icons_and_images/custom_logo.dart';
 import 'package:flutter_blog/ui/widgets/button_items/button/custom_text_button.dart';
 import 'package:flutter_blog/ui/widgets/custom_nav_appbar.dart';
 import 'package:flutter_blog/ui/widgets/line/custom_line_bold.dart';
