+++ lib/ui/widgets/custom_modal_bottom_sheet.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class CustomModalBottomSheet extends StatelessWidget {
+  const CustomModalBottomSheet({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
