+++ lib/ui/widgets/icons_and_images/custom_review_icon.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_blog/ui/widgets/custom_star_icon.dart';
+import 'package:flutter_blog/ui/widgets/icons_and_images/custom_star_icon.dart';
 
 class CustomReviewIcon extends StatelessWidget {
   const CustomReviewIcon({
