+++ lib/ui/widgets/icons_and_images/custom_star_icon.dart
