+++ lib/ui/widgets/icons_and_images/custom_svg_image.dart
